refactor: view 기능을 담은 파일 이름명 수정
