+++ src/js/view/NewsTitle.js
