+++ src/js/view/Notification.js
